Refactor telemetry files
